+++ src/Model/Image.php
@@ -18,7 +18,7 @@ class Image extends Simple
      */
     public function getPath()
     {
-        return (string)$this->xml;
+        return trim((string)$this->xml);
     }
 
     /**
@@ -27,7 +27,7 @@ class Image extends Simple
     public function getCaption()
     {
         if ($xml = $this->xpath("//c:ЗначениеРеквизита[contains(c:Значение, '{$this->path}#')]")) {
-            return (string)\array_slice(explode('#', (string)$xml[0]->Значение), 1);
+            return (string)current(\array_slice(explode('#', (string)$xml[0]->Значение), 1));
         }
 
         return '';