+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # 0.2.4
+* Переименован метод \Zenwalker\CommerceML\Model\Classifier::getReferenceBookValue => getReferenceBookValueById
+* Переименован метод defaultProperties в propertyAliases в базовом классе модели
 * Добавлены тесты
 * Добавлен метод getOffers у продукта, возращает все предложения, метод getOffer помечен как устаревший, будет исключен в 0.3.0
 * Добавлен метод getOffersById у offerPackage, возращает все предложения, метод getOfferById помечен как устаревший, будет исключен в 0.3.0